# `find_first_of` vectorized: generalize fast approach for 4 and 8 bytes elements