@AlexGuteniev 2024-04-23
There are unexpected changes in benchmark results in other lines (presumably due to different loop alignments), but the target for this optimization are the following, where the needle is too big to fit AVX2 reg:

Before:
```
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
bm<uint32_t>/1011/11       1035 ns          643 ns      2357895
bm<uint64_t>/1011/11       1371 ns         1029 ns      1518644
bm<uint64_t>/3056/7        3183 ns         2066 ns       597333
```
After:
```
---------------------------------------------------------------
Benchmark                     Time             CPU   Iterations
---------------------------------------------------------------
bm<uint32_t>/1011/11        329 ns          195 ns      8960000
bm<uint64_t>/1011/11        636 ns          289 ns      3733333
bm<uint64_t>/3056/7        1217 ns          330 ns      8960000
```